requirements.txt
openai~=0.28.0
requests~=2.32.3
flask~=3.1.1
python-dotenv~=1.1.0
scikit-learn==1.3.0
numpy==1.24.3